% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_cmprsk_regr_finegrayhelper.R
\name{configure_task}
\alias{configure_task}
\title{Configure a task for competing risks analysis}
\usage{
configure_task(task_name = "pbc", stratum = NULL, features = "trt")
}
\arguments{
\item{stratum}{Character name of the stratum column.}

\item{features}{Character vector of feature names.}
}
\value{
An mlr3 task.
}
\description{
Configure a task for competing risks analysis
}
\keyword{internal}
