% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{skip_if_not_installed}
\alias{skip_if_not_installed}
\title{A helper function to skip a test in \code{testthat} if a specified R package is not installed.
It checks if the package is available using \code{requireNamespace()} and skips the test with
an informative message if the package is not found.}
\usage{
skip_if_not_installed(pkg)
}
\arguments{
\item{pkg}{\code{character(1)} \cr
The name of the package to check for installation.}
}
\value{
Invisible \code{NULL}. The function is called for its side effect of skipping the test.
}
\description{
A helper function to skip a test in \code{testthat} if a specified R package is not installed.
It checks if the package is available using \code{requireNamespace()} and skips the test with
an informative message if the package is not found.
}
\keyword{internal}
