% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_cmprsk_regr_finegrayhelper.R
\name{create_partition}
\alias{create_partition}
\title{Create a train-test partition for a task}
\usage{
create_partition(task, ratio = 0.7)
}
\arguments{
\item{task}{An mlr3 task.}
}
\value{
A list with train and test row IDs.
}
\description{
Create a train-test partition for a task
}
\keyword{internal}
